requirements.txt
python-dotenv>=1.0.0
flask
flask_cors
sqlite3
openai
beautifulsoup4